Adds more matchers and changes modules to match deno coding style
