+++ reactive/memo.ts
@@ -1,5 +1,4 @@
 // Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
-// This module is browser compatible.
 
 /**
  * Contains the implementation for the {@link memoSignal} function.
@@ -28,8 +27,8 @@ import {
   markAsSignal,
   type MemoizedSignal,
   type MemoizedSignalOptions,
-} from './api.ts';
-import { ReactiveNode } from './reactive_node.ts';
+} from './_api.ts';
+import { ReactiveNode } from './_reactive_node.ts';
 import { untrackedSignal } from './untracked.ts';
 
 /**
