+++ testing/_collect.ts
@@ -1,2 +1,1 @@
 // Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
-// This module is browser compatible.
