+++ testing/_default_config.ts
@@ -1,3 +1,1 @@
 // Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
-
-export * from './_type_awaitable.ts';
