+++ testing/_matchers/_matcher_impl_options.ts
@@ -1,5 +1,4 @@
 // Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
-// This module is browser compatible.
 
 import type { ExpectContext } from './_context.ts';
 
@@ -7,6 +6,7 @@ import type { ExpectContext } from './_context.ts';
  * Options passed to the matcher implementation functions.
  *
  * @template T The expected type of the value.
+ * @internal
  */
 export interface MatcherImplOptions<T = unknown> {
   /** The expected value. */
