+++ testing/mod.ts
@@ -1,7 +1,6 @@
 // Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
-// This module is browser compatible.
 
-export { expect } from './expect/mod.ts';
-export { group } from './group.ts';
-export { afterAll, afterEach, beforeAll, beforeEach } from './hooks.ts';
-export { test } from './test.ts';
+export { expect } from './_expect.ts';
+export { group } from './_group.ts';
+export { afterAll, afterEach, beforeAll, beforeEach } from './_hooks.ts';
+export { test } from './_test.ts';
