+++ utils/_type_awaitable.ts
@@ -1,3 +1,5 @@
+// Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
+
 /**
  * Type for a value that can be awaited.
  *
